Fix for issue #14. Better PIC CAV error checking and some state-machine bug-fixes
The solution for #14 (that should work with most players) is for the CAV
PIC capture state-machine to check if the player is paused during
capture.  If the player pauses but the current frame isn't the target
end-frame, the state-machine restarts the play-back and continues
capturing.

There were also a few minor bugs around handling automatic capture at
the end of discs.  These have now been (hopefully) fixed.