Several bug fixes
Fix for issue #56
Changed threaded code to exit faster on quit (added 'stop' function to
player control and usb-device threads and added commands to the class
deconstuctor for the main window).
Removed 5 retries from the player communication (as the player control
loop will take care of this)
Sped up some of the thread sleep commands for faster response to quit.