Fix a race condition when writing data to disk.
When the code that's reading from the data buffers finds an empty
buffer, it does usleep(100) then continues with the *next* buffer. This
can go wrong: usleep's argument is a *minimum* time to sleep, so if the
operating system delays it for longer then this can happen:

* Disk thread checks buffer 2; it's empty, sleep
* Capture thread fills buffer 2
* Capture thread fills buffer 3
* Disk thread checks buffer 3; it's full, write it
* Disk thread checks buffer 0; it's empty, sleep
* Disk thread checks buffer 1; it's empty, sleep
* Disk thread checks buffer 2; it's full, write it

So the output's been written in the wrong order.

To fix this, prevent the disk thread from moving on to the next buffer
if it hasn't found any data and the capture is still running.