Check for errors when writing to disk.
The code already checks that opening the file was successful, but not
that data was actually written -- e.g. if the disk fills up or is
disconnected during capture.

Because we're using QFile in buffered mode, we check the return value
from write() and we add an explicit flush() to write out the rest of the
buffer before closing. (Oddly, QFile::close doesn't provide any way to
detect errors from the underlying close().)