Make the test data count 0-1020 rather than 0-1023.
The sequence was previously 1024 values long, which is a divisor of the
transfer and disk buffer lengths, so every disk buffer had the same
contents. It's now 1021 values long (a prime), so it has no common
factors with any of the other lengths -- which means we're more likely
to detect problems that involve dropping or reordering transfers or disk
buffers.

Make the test sequence verification code in DomesdayDuplicator and
dddutil detect the sequence length automatically the first time it
wraps.