Add sequence numbers to the sample values.
We transfer 10-bit samples in 16-bit words, so 6 bits were previously
unused. This commit adds a sequence number to the unused bits when they
are sampled, which makes it possible for DomesdayDuplicator to detect
lost samples (at any stage of the process) during capture.

Most of the changes to the Verilog are to make the FIFOs and
intermediate wiring 16 bits wide. We then add a 22-bit sample counter to
dataGenerator that works in the same way as the test data counter; the
bottom 16 bits are ignored, and the top 6 bits give us the sequence
number.

The sequence numbers count from 0-62, rather than 0-63, to avoid having
common factors with buffer sizes.

This commit also changes DomesdayDuplicator to strip off the sequence
numbers before decoding the samples (verification will come next).