Rework how amplitude buffers are captured.
Previously, when an amplitude display was enabled, the UsbCapture thread
converted every disk buffer into 16-bit form and copied it to a
temporary buffer, which was very expensive, and meant it had to know
whether the display was enabled.

Now it's done on request: the static function that returns a disk buffer
gives back the most recently-captured one, and sets a flag saying that a
new one should be captured. So overall the UsbCapture thread will only
convert buffers at the rate the display is updating.

The expensive part of the conversion now happens in the main thread, so
it should interfere less with the capture thread. This also considerably
simplifies the conversion code -- there's no need to use QtMultimedia,
which makes it compatible with Qt 5 again.